src/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
load-dotenv==0.1.0
lxml==4.9.3
numpy==1.25.1
opencv-contrib-python==4.8.0.74
outcome==1.2.0
packaging==23.1
Pillow==10.0.0
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
slack-sdk==3.21.3
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
